ReCell
The rising potential of this comparatively under-the-radar market fuels the need for an ML-based solution to develop a dynamic pricing strategy for used and refurbished devices. ReCell, a startup aiming to tap the potential in this market. As a data scientist, analyzed the data provided and built a linear regression model to predict the price of a used phone/tablet and identified factors that significantly influence it.